LayoutTests/http/tests/from-origin/image-from-origin-same-blocked-expected.txt

CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
CONSOLE MESSAGE: Cannot load image http://localhost:8000/from-origin/resources/image.php?fromOrigin=same due to access control checks.
Tests that a cross-origin image load fails if the server blocks it with a 'From-Origin: same' response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Image load blocked.
PASS successfullyParsed is true

TEST COMPLETE
